chore: reconfigure dirs, lift main -> cli.py
