+++ cli.py
@@ -3,16 +3,13 @@ from typing import Optional
 import numpy as np
 import typer
 
-from ctmds1.generate_random_prices import generate_random_prices
+from utils.generate_random_prices import generate_random_prices
+from utils.validation import validate_non_negative
 
+app = typer.Typer()
 
-def validate_non_negative(value: int) -> int:
-    if value < 0:
-        raise typer.BadParameter("Number must be non-negative")
-    return value
-
-
-def main(
+@app.command()
+def random_prices(
     num: int = typer.Argument(
         ..., help="Number of random prices to generate", callback=validate_non_negative
     ),
@@ -22,5 +19,9 @@ def main(
     return result
 
 
+# @app.command()
+# def daily_prices()
+#     # todo
+
 if __name__ == "__main__":
-    typer.run(main)
+    app()
