+++ tests/test_random_prices_print.py
@@ -4,40 +4,40 @@ import numpy as np
 import pytest
 import typer
 
-from ctmds1.main import main
+from cli import random_prices
 
 
-def test_main_with_invalid_arg(monkeypatch, capsys):
+def test_random_prices_with_invalid_arg(monkeypatch, capsys):
     monkeypatch.setattr(sys, "argv", ["prog", "abc"])
     with pytest.raises(SystemExit) as exc_info:
-        typer.run(main)
+        typer.run(random_prices)
     assert exc_info.value.code == 2
     captured = capsys.readouterr()
     assert "Invalid value for 'NUM': 'abc' is not a valid integer." in captured.err
 
 
-def test_main_with_invalid_negative_arg(monkeypatch, capsys):
+def test_random_prices_with_invalid_negative_arg(monkeypatch, capsys):
     monkeypatch.setattr(sys, "argv", ["prog", "-1"])
     with pytest.raises(SystemExit) as exc_info:
-        typer.run(main)
+        typer.run(random_prices)
     assert exc_info.value.code == 2
     captured = capsys.readouterr()
     assert "No such option: -1" in captured.err
 
 
-def test_main_with_invalid_missing_arg(monkeypatch, capsys):
+def test_random_prices_with_invalid_missing_arg(monkeypatch, capsys):
     monkeypatch.setattr(sys, "argv", ["prog"])
     with pytest.raises(SystemExit) as exc_info:
-        typer.run(main)
+        typer.run(random_prices)
     assert exc_info.value.code == 2
     captured = capsys.readouterr()
     assert "Missing argument 'NUM'. " in captured.err
 
 
-def test_main_with_valid_arg(monkeypatch, capsys):
+def test_random_prices_with_valid_arg(monkeypatch, capsys):
     monkeypatch.setattr(sys, "argv", ["prog", "3"])
     with pytest.raises(SystemExit) as exc_info:
-        typer.run(main)
+        typer.run(random_prices)
     # success
     assert exc_info.value.code == 0
     captured = capsys.readouterr()
@@ -48,10 +48,10 @@ def test_main_with_valid_arg(monkeypatch, capsys):
     assert all(0 <= x <= 100 for x in array)  # Check range if that's your requirement
 
 
-def test_main_with_valid_zero_arg(monkeypatch, capsys):
+def test_random_prices_with_valid_zero_arg(monkeypatch, capsys):
     monkeypatch.setattr(sys, "argv", ["prog", "0"])
     with pytest.raises(SystemExit) as exc_info:
-        typer.run(main)
+        typer.run(random_prices)
     # success
     assert exc_info.value.code == 0
     captured = capsys.readouterr()
